Add AdminClaim it tests
